requirements.txt
telethon==1.36.0
asyncpg==0.29.0
python-decouple==3.8
